Move Reconnect DevTools toolbar button to own module
