+++ front_end/entrypoints/rn_fusebox/BUILD.gn
@@ -9,6 +9,7 @@ import("../visibility.gni")
 
 devtools_module("rn_fusebox") {
   sources = [
+    "FuseboxReconnectDeviceButton.ts",
     "FuseboxProfilingBuildObserver.ts",
   ]
